Make smaller
We've stopped having one script serve as both the node version and the
browser version.  Now we have one browser version and then a node script
that uses the browser version but fixes the global variable being set.

Since node already has a lot of this functionallity built in it doesn't
make sense to hurt our size for something that probably won't be used
too often.  And this doesn't actually change the end result.